+++ models/gambas_model.py
