+++ models/mamba_modules3D.py
@@ -12,17 +12,15 @@ from os.path import join as pjoin
 import torch
 import torch.nn as nn
 import numpy as np
-from torch.nn import CrossEntropyLoss, Dropout, Softmax, Linear, Conv2d, LayerNorm
-from torch.nn.modules.utils import _pair
+from torch.nn import CrossEntropyLoss, Dropout, Softmax, Linear, LayerNorm
 import torch.nn.functional as F
-from scipy import ndimage
-from . import transformer_configs as configs
-from path_generate import *
+from .path_generate import generate3d, gilbert3d, generate_gilbert_indices_3D
 
 from mamba_ssm import Mamba
 
 
-# Define a resnet block
+logger = logging.getLogger(__name__)
+
 class ResnetBlock(nn.Module):
     def __init__(self, dim, padding_type, norm_layer, use_dropout, use_bias,dim2=None):
         super(ResnetBlock, self).__init__()
@@ -65,15 +63,12 @@ class ResnetBlock(nn.Module):
         out = x + self.conv_block(x)
         return out
 
-
-# Mamba version
 class BottleneckCNN(nn.Module):
-    def __init__(self, config):
+    def __init__(self):
         super(BottleneckCNN, self).__init__()
-        self.config = config
-        use_bias = False
+        use_bias = True
         norm_layer = nn.InstanceNorm3d
-        padding_type = "reflect"
+        padding_type = 'replicate'
         
         # Residual CNN
         model = [ResnetBlock(256, padding_type=padding_type, norm_layer=norm_layer, use_dropout=False,
@@ -94,32 +89,51 @@ class MambaLayer(nn.Module):
         expand (int): Block expansion factor.
     
     """
-    def __init__(self, dim, d_state=16, d_conv=4, expand=2):
+    def __init__(self, dim, d_state=16, d_conv=4, expand=2): # Before it was d_state=16, d_conv=4, expand=2
         super().__init__()
         self.dim = dim
         self.norm = nn.LayerNorm(dim)
-        self.mamba = Mamba(d_model=dim, d_state=d_state, d_conv=d_conv, expand=expand)
+        self.mamba1 = Mamba(d_model=dim, d_state=d_state, d_conv=d_conv, expand=expand)
+        self.mamba2 = Mamba(d_model=dim, d_state=d_state, d_conv=d_conv, expand=expand)
+
+        self.conv1d = nn.Conv3d(in_channels=512, out_channels=256, kernel_size=1)
+        self.generator = gilbert3d(32, 32, 32)
+        self.gilbert_indices = generate_gilbert_indices_3D(32, 32, 32, self.generator).expand(-1, dim, -1).permute(0, 2, 1) # Before it was 64, 64, 8
+        self.degilbert_indices = torch.argsort(self.gilbert_indices)
+        self.gilbert_r_indices = torch.flip(self.gilbert_indices, dims=[2])
+        self.degilbert_r_indices = torch.argsort(self.gilbert_r_indices)
     
     def forward(self, x):
         B, C, D, H, W = x.shape
 
         # Check model dimension
         assert C == self.dim
-        
-        # Convert input from (B, C, H, W) to (B, H*W, C)
-        x = x.float().view(B, C, -1).permute(0, 2, 1)
-        
-        # Normalize and pass through Mamba layer
-        norm_out = self.norm(x)
-        mamba_out = self.mamba(norm_out)
 
-        # Convert output from (B, H*W, C) to (B, C, H, W)
-        out = mamba_out.permute(0, 2, 1).view(B, C, D, H, W)
+        # Bidirectional mamba
+        x1 = x.view(B, C, -1).permute(0, 2, 1)
+        device = 'cuda:0'
+        self.gilbert_indices = self.gilbert_indices.to(device)
+        x1 = torch.gather(x1, 1, self.gilbert_indices)
+        x2 = torch.flip(x1, dims=[1])
 
-        return out
+        # Pass forwad and reverse through mamba
+        norm_out1 = self.norm(x1)
+        mamba_out1 = self.mamba1(norm_out1)
+        norm_out2 = self.norm(x2)
+        mamba_out2 = self.mamba2(norm_out2)
 
-class cmMambaWithCNN(nn.Module):
-    """ Channel-mixed Mamba (cmMamba) block with residual CNN block
+        self.degilbert_indices = self.degilbert_indices.to(device)
+        self.degilbert_r_indices = self.degilbert_r_indices.to(device)
+        out1 = torch.gather(mamba_out1, 1, self.degilbert_indices).permute(0, 2, 1).view(B, C, D, H, W)
+        out2 = torch.gather(mamba_out2, 1, self.degilbert_r_indices).permute(0, 2, 1).view(B, C, D, H, W)
+
+        concatenated = torch.cat((out1, out2), dim=1)
+        output = self.conv1d(concatenated)
+
+        return output
+
+class ccMambaWithCNN(nn.Module):
+    """ Channel-compressed Mamba (ccMamba) block with residual CNN block
 
     Args:
         config (dict): Model configuration.
@@ -134,17 +148,16 @@ class cmMambaWithCNN(nn.Module):
         img_size (int): Image size.
     
     """
-    def __init__(self, config, in_channels, d_state=16, d_conv=4, expand=2, ngf=64, norm_layer=nn.BatchNorm2d, use_bias=False):
+    def __init__(self, in_channels, d_state=16, d_conv=4, expand=2, ngf=64, norm_layer=nn.BatchNorm2d, use_bias=True):
         super().__init__()
         # Mamba block
         self.mamba_layer = MambaLayer(
             dim=in_channels, d_state=d_state, d_conv=d_conv, expand=expand
         )
 
-        self.config = config
         ngf = 64
-        padding_type = "reflect"
-        use_bias = False
+        padding_type = 'replicate'
+        use_bias = True
         norm_layer = nn.InstanceNorm3d
 
         # Channel compression block
@@ -168,20 +181,21 @@ class cmMambaWithCNN(nn.Module):
 
         return x
 
-class I2IMamba(nn.Module):
-    def __init__(self, config, input_dim, img_size=224, output_dim=3):
-        super(I2IMamba, self).__init__()
-        self.config = config
+########Generator############
+class GAMBAS(nn.Module):
+    def __init__(self, input_dim, img_size=224, output_dim=3):
+        super(GAMBAS, self).__init__()
+        # self.config = config
         output_nc = output_dim
         ngf = 64
-        use_bias = False
+        use_bias = True
         norm_layer = nn.InstanceNorm3d
-        padding_type = "reflect"
+        padding_type = "replication"
         mult = 4
 
         ############################################################################################
         # Layer1-Encoder1
-        model = [nn.ReflectionPad3d(3),
+        model = [nn.ReplicationPad3d(3),
                  nn.Conv3d(input_dim, ngf, kernel_size=7, padding=0, 
                            bias=use_bias),
                  norm_layer(ngf),
@@ -211,34 +225,31 @@ class I2IMamba(nn.Module):
                  nn.ReLU(True)]
         
         setattr(self, "encoder_3", nn.Sequential(*model))
-        ####################################ART Blocks##############################################
+        ############################################################################################
+        # Bottlenck Layers
         mult = 4
         img_size = 256 
         input_dim = 256 # Adjust this according to new input dimension
 
-        # Episodic bottleneck
-        # cmMamba block with residual CNN block
-        self.bottleneck_1 = cmMambaWithCNN(self.config, input_dim)
-        # self.bottleneck_1 = BottleneckCNN(self.config)
+        # ccMamba block with residual CNN block
+        self.bottleneck_1 = ccMambaWithCNN(input_dim)
         
-        self.bottleneck_2 = BottleneckCNN(self.config)
-        self.bottleneck_3 = BottleneckCNN(self.config)
-        self.bottleneck_4 = BottleneckCNN(self.config)
+        self.bottleneck_2 = BottleneckCNN()
+        self.bottleneck_3 = BottleneckCNN()
+        self.bottleneck_4 = BottleneckCNN()
 
-        # cmMamba block with residual CNN block
-        self.bottleneck_5 = cmMambaWithCNN(self.config, input_dim)
-        # self.bottleneck_5 = BottleneckCNN(self.config)
+        # ccMamba block with residual CNN block
+        self.bottleneck_5 = ccMambaWithCNN(input_dim)
         
-        self.bottleneck_6 = BottleneckCNN(self.config)
-        self.bottleneck_7 = BottleneckCNN(self.config)
-        self.bottleneck_8 = BottleneckCNN(self.config)
+        self.bottleneck_6 = BottleneckCNN()
+        self.bottleneck_7 = BottleneckCNN()
+        self.bottleneck_8 = BottleneckCNN()
 
-        # cmMamba block with residual CNN block
-        self.bottleneck_9 = cmMambaWithCNN(self.config, input_dim)
-        # self.bottleneck_9 = BottleneckCNN(self.config)
+        # ccMamba block with residual CNN block
+        self.bottleneck_9 = ccMambaWithCNN(input_dim)
 
         ############################################################################################
-        # Layer13-Decoder1 - currently removed the additional in_channels (removed * 2 for first argument), taking away skip connection to here
+        # Layer13-Decoder1
         n_downsampling = 2
         i = 0
         mult = 2 ** (n_downsampling - i)
@@ -265,7 +276,7 @@ class I2IMamba(nn.Module):
         ############################################################################################
         # Layer15-Decoder3
         model = []
-        model = [nn.ReflectionPad3d(3)]
+        model = [nn.ReplicationPad3d(3)]
         model += [nn.Conv3d(ngf, output_dim, kernel_size=7, padding=0)]
         model += [nn.Tanh()]
         setattr(self, "decoder_3", nn.Sequential(*model))
@@ -288,13 +299,11 @@ class I2IMamba(nn.Module):
         x = self.bottleneck_9(x)
 
         # Decoder
-        # x = self.decoder_1(torch.cat([x, x3], dim=1))
-        # x = self.decoder_2(torch.cat([x, x2], dim=1))
-        # x = self.decoder_3(torch.cat([x, x1], dim=1))
         x = self.decoder_1(x)
         x = self.decoder_2(x)
         x = self.decoder_3(x)
         return x
+
 
 class channel_compression(nn.Module):
     def __init__(self, in_channels, out_channels, stride=1):
@@ -310,16 +319,16 @@ class channel_compression(nn.Module):
 
         if stride != 1 or in_channels != out_channels:
           self.skip = nn.Sequential(
-            nn.Conv3d(in_channels=in_channels, out_channels=out_channels, kernel_size=1, stride=stride, bias=False),
+            nn.Conv3d(in_channels=in_channels, out_channels=out_channels, kernel_size=1, stride=stride, bias=True),
             nn.InstanceNorm3d(out_channels))
         else:
           self.skip = None
 
         self.block = nn.Sequential(
-            nn.Conv3d(in_channels=in_channels, out_channels=out_channels, kernel_size=3, padding=1, stride=1, bias=False),
+            nn.Conv3d(in_channels=in_channels, out_channels=out_channels, kernel_size=3, padding=1, stride=1, bias=True),
             nn.InstanceNorm3d(out_channels),
             nn.ReLU(),
-            nn.Conv3d(in_channels=out_channels, out_channels=out_channels, kernel_size=3, padding=1, stride=1, bias=False),
+            nn.Conv3d(in_channels=out_channels, out_channels=out_channels, kernel_size=3, padding=1, stride=1, bias=True),
             nn.InstanceNorm3d(out_channels))
 
     def forward(self, x):
@@ -327,5 +336,3 @@ class channel_compression(nn.Module):
         out += (x if self.skip is None else self.skip(x))
         out = F.relu(out)
         return out
-
-
